### 2023-04-18 #5385 @editorialbot
[PRE REVIEW]: Effluent - A python package for modelling effluent discharge
**Submitting author:** @pnsaevik (<a href="http://orcid.org/0000-0002-7301-2008">Pål Næverlid Sævik</a>)
**Repository:** https://github.com/pnsaevik/effluent
**Branch with paper.md** (empty if default branch):
**Version:** v1.2
