### 2023-04-26 #5385 @pnsaevik
In addition to the reviewers in your pool, you might want to contact the github user `nordam` as a potential reviewer. Apart from this, users which have already agreed to be JOSS reviewers and have Python & oceanography/fluid dynamics and/or turbulence modelling in their listed preference, might be suitable candidates.
