### 2023-05-30 #5385 @kthyng
@tomchor and @castelao Do you have the time and interest in reviewing this JOSS submission? As you know, reviews are done in a github issue and consist of a thorough software review. I would ask you to get started with your review in the next 3 weeks or so since usually the process is iterative between you and the author. You can check out the [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Thank you for your consideration!
