### 2023-06-01 #5385 @castelao
@kthyng , JOSS used to have a rubric for review with an explicit list of checkmarks. Do you still use it? Where may I find it? Thanks,
