### 2023-06-07 8bd0c0d
Remove dask from test prerequisites
This is done to verify that the CI tests fail
