+++ setup.cfg
@@ -31,6 +31,7 @@ install_requires =
     netcdf4
     holoviews
     scipy
+    dask
     tomli
 python_requires = >=3.6
 
