### 2023-06-08 e782bfa
Add command line test to circleCI script
