+++ .circleci/config.yml
@@ -23,6 +23,13 @@ jobs:
           name: "Install package"
           command: pip install .
       - run:
+          name: "Run simple command-line example"
+          command: |
+            mkdir ./cli_test/
+            cp ./docs/source/examples/jet/config.toml ./cli_test/
+            python -m effluent ./cli_test/config.toml
+            cat ./cli_test/out.csv
+      - run:
           name: "Run pytest"
           command: pytest
       - run:
