### 2023-06-08 f512cd6
Edit CLI test path
