+++ setup.cfg
@@ -38,6 +38,10 @@ python_requires = >=3.6
 [options.packages.find]
 where=src
 
+[options.entry_points]
+console_scripts =
+    effluent = effluent::main
+
 [options.package_data]
 * =
     *.toml
