+++ setup.cfg
@@ -40,7 +40,7 @@ where=src
 
 [options.entry_points]
 console_scripts =
-    effluent = effluent::main
+    effluent = effluent:main
 
 [options.package_data]
 * =
