### 2023-06-08 b3fae1e
Use argparse in main script
