+++ docs/source/conf.py
@@ -60,6 +60,8 @@ release = getversion()
 extensions = [
     'sphinx.ext.mathjax',
     'sphinx.ext.doctest',
+    'sphinx.ext.autodoc',
+    'autoapi.extension',
     'matplotlib.sphinxext.plot_directive',
 ]
 
@@ -94,3 +96,11 @@ html_theme = 'alabaster'
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
+
+
+# -- Options for Sphinx AutoAPI -----------------------------------------------
+
+autoapi_dirs = ['../../src']
+autoapi_add_toctree_entry = True
+autoapi_member_order = 'alphabetical'
+autoapi_template_dir = '_templates/autoapi'
