+++ src/effluent/__init__.py
@@ -1,3 +1,7 @@
+"""
+This is a package docstring.
+"""
+
 from .script import run, main
 
 __version__ = "1.2"
