+++ docs/source/conf.py
@@ -101,7 +101,7 @@ html_static_path = ['_static']
 # -- Options for Sphinx AutoAPI -----------------------------------------------
 
 autoapi_dirs = ['../../src']
-autoapi_add_toctree_entry = True
+# autoapi_add_toctree_entry = True
 autoapi_member_order = 'alphabetical'
 autoapi_template_dir = '_templates/autoapi'
 autoapi_options = [
