### 2023-06-09 0c53fc8
Add autoapi extension to circleci builds
