### 2023-06-09 fb210c2
Add autoapi extension to readthedocs builds
