+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Command line script for starting the program
+- Citation guidance
 
 ### Fixed
 - Included dask as an explicit dependency
