### 2023-06-12 cdd6d61
Remove main method from package root
