+++ src/effluent/__init__.py
@@ -1,7 +1,8 @@
 """
-This is a package docstring.
+The main package contains only a single function, :func:`effluent.run`, which is the
+main entry point of the application.
 """
 
-from .script import run, main
+from .script import run
 
 __version__ = "1.2"
