### 2023-06-12 6bd858a
Add documentation to the model subpackage
