+++ src/effluent/__init__.py
@@ -6,3 +6,6 @@ main entry point of the application.
 from .script import run
 
 __version__ = "1.2"
+"""
+The version number
+"""
