+++ src/effluent/io.py
@@ -1,5 +1,5 @@
 """
-This is the io package
+The package contains functions and classes for reading and writing simulation data.
 """
 
 import abc
@@ -8,7 +8,13 @@ import numpy as np
 
 class Pipe:
     """
-    This is the Pipe class
+    Data about the pipe and the effluent release.
+
+    The constructor takes an xarray.Dataset object as input. If the source data is
+    in the form of a data file, the factory method should be used instead.
+
+    :param dset: An xarray.Dataset object with variables ``depth``, ``u``,
+        ``w``, ``dens`` and ``diam``, all indexed by the coordinate ``time``.
     """
 
     def __init__(self, dset):
@@ -18,6 +24,12 @@ class Pipe:
 
     @staticmethod
     def from_config(conf):
+        """
+        Initialize using :doc:`configuration parameters </config/pipe>`
+
+        :param conf: A dict of configuration parameters
+        :return: An initialized object
+        """
         if 'csv' in conf:
             return Pipe.from_csv_file(**conf['csv'])
         elif 'nc' in conf:
@@ -89,6 +101,14 @@ class Pipe:
         return Pipe.from_dataframe(df)
 
     def select(self, time):
+        """
+        Interpolate pipe parameters to a specific point in time
+
+        :param time: A time in numpy.datetime64 format
+        :return: An xarray.Dataset object with variables ``depth``, ``u``, ``w``,
+            ``dens`` and ``diam``.
+        """
+
         if self._dset.dims['time'] == 1:
             # No interpolation is possible if there is only 1 time entry
             return self._dset.isel(time=0)
@@ -98,6 +118,12 @@ class Pipe:
 
 
 def read_csv(file):
+    """
+    Read csv file, and return a pandas.DataFrame
+
+    :param file: File name
+    :return: A pandas.DataFrame object
+    """
     import pandas as pd
 
     return pd.read_csv(
@@ -113,15 +139,32 @@ def read_csv(file):
 
 class Ambient:
     """
-    This is the Ambient class
+    Data about the ambient ocean.
+
+    This is an abstract base class with no explicit constructor. To initialize an
+    instance of the class, use the factory method.
     """
 
     @abc.abstractmethod
     def select(self, time):
+        """
+        Compute the ambient conditions at a specific time.
+
+        :param time: A numpy.datetime64 object
+        :return: An xarray.Dataset object with variables ``u``, ``v`` and ``dens``, all
+            indexed by the coordinate ``depth``.
+        """
         return NotImplementedError
 
     @staticmethod
     def from_config(conf):
+        """
+        Initialize using :doc:`configuration parameters </config/ambient>`
+
+        :param conf: A dict of configuration parameters
+        :return: An initialized object
+        """
+
         if 'csv' in conf:
             return Ambient.from_csv_file(**conf['csv'])
         elif 'nc' in conf:
@@ -191,10 +234,23 @@ class Ambient:
         return Ambient.from_dataset(dset)
 
     def close(self):
+        """
+        Close the underlying data source
+        """
         pass
 
 
 class AmbientXarray(Ambient):
+    """
+    Data about the ambient ocean, from in-memory dataset.
+
+    This subclass is using an xarray.Dataset object as its data source. The dataset
+    should have variables ``u``, ``v`` and ``dens``, all indexed by the coordinates
+    ``depth`` and ``time``.
+
+    :param dset: An xarray.Dataset object
+    """
+
     def __init__(self, dset):
         self._dset = dset
         self._tmin = dset.time[0].values
@@ -211,11 +267,21 @@ class AmbientXarray(Ambient):
 
 class Output:
     """
-    This is the Output class
+    Class for writing simulation output to disk
+
[…]
