### 2023-06-12 cdc3c2d
Document the roms module
