### 2023-06-12 4d83252
Document the solver module
