### 2023-06-13 #5385 @kthyng
@bastorer I don't suppose you're up for a JOSS review so soon after your epic review ended? If you are, we'd love to have you review this submission!  You probably got a feel for in being on the other side,  but here are some docs you can check out too: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
