### 2023-06-13 #5554 @editorialbot
[REVIEW]: Effluent - A python package for modelling effluent discharge
**Submitting author:** @pnsaevik (<a href="http://orcid.org/0000-0002-7301-2008">Pål Næverlid Sævik</a>)
**Repository:** https://github.com/pnsaevik/effluent
**Branch with paper.md** (empty if default branch):
**Version:** 1.3
