### 2023-06-14 aa6c6f1
Restructure internal solver methods
