### 2023-06-14 56d79c0
Factor out common code
