### 2023-06-15 5ba809c
Turn sphinx warnings into errors
Otherwise, examples may break without anyone knowing
