### 2023-06-15 2d8e0cf
Add extra sphinx build step
The previous commit did not work as expected
