+++ .circleci/config.yml
@@ -31,8 +31,11 @@ jobs:
           name: "Run pytest"
           command: pytest
       - run:
+          name: "Run sphinx html build"
+          command: sphinx-build -b html -W --keep-going -n ./docs/source/ ./docs/build/
+      - run:
           name: "Run sphinx doctest"
-          command: sphinx-build -M doctest -W --keep-going -n ./docs/source/ ./docs/build/
+          command: sphinx-build -M doctest ./docs/source/ ./docs/build/
       - run:
           name: "Run sphinx linkcheck"
           command: sphinx-build -M linkcheck ./docs/source/ ./docs/build/
