### 2023-06-15 5c3609d
Update example
Function signature changed
