+++ .circleci/config.yml
@@ -32,7 +32,7 @@ jobs:
           command: pytest
       - run:
           name: "Run sphinx html build"
-          command: sphinx-build -b html -W --keep-going -n ./docs/source/ ./docs/build/
+          command: sphinx-build -v -b html -W --keep-going -n ./docs/source/ ./docs/build/
       - run:
           name: "Run sphinx doctest"
           command: sphinx-build -M doctest ./docs/source/ ./docs/build/
