### 2023-06-15 43f9d00
Re-introduce type hints
