+++ docs/source/externals.rst
@@ -1,0 +1,12 @@
+:orphan:
+
+=========================
+External documentation
+=========================
+
+.. py:class:: netCDF4.Dataset
+
+    A netCDF Dataset object is a collection of dimensions, groups, variables
+    and attributes. Together they describe the meaning of data and relations
+    among data fields stored in a netCDF file. Full documentation is available
+    `here <https://unidata.github.io/netcdf4-python/>`_.
