### 2023-06-15 24817bb
Add type hints
