### 2023-06-22 ed26bdb
Add references to software paper
