+++ docs/paper.md
@@ -37,15 +37,15 @@ impact.
 
 `effluent` is an open-source python package for simulating the dispersion of
 effluent discharges from wastewater pipes. The underlying model is based on
-`@lee:2003`. The implementation also contains couplings to the popular
-open-source ocean model ROMS `[@shchepetkin:2005]`.
+[@Lee2003]. The implementation also contains couplings to the popular
+open-source ocean model ROMS [@Shchepetkin2005].
 
 There already exists multiple closed-source applications for modelling outfall
-dispersion, such as CORMIX `[@jirka:2004]`, VISUAL PLUMES `[@frick:2004]`
-or VISJET `[@lee:2003]`. These are well-suited for governmental and industrial
+dispersion, such as CORMIX [@Jirka2004], VISUAL PLUMES [@Frick2004]
+or VISJET [@Lee2003]. These are well-suited for governmental and industrial
 applications. In research, however, it is desirable to have an open-source,
 transparent implementation which can also be adapted and tailored to the
-specific needs of a reserach project. The package `effluent` provides this,
+specific needs of a research project. The package `effluent` provides this,
 along with a detailed online documentation with usage examples.
 
 # Acknowledgements
@@ -53,3 +53,118 @@ along with a detailed online documentation with usage examples.
 This work is financed by the Institute of Marine Sciences, Norway.
 
 # References
+
+---
+references:
+- id: Frick2004
+  author:
+    - family: Frick
+      given: Walter E.
+  citation-key: Frick2004
+  accessed:
+    - year: 2021
+      month: 2
+      day: 19
+  collection-title: Facilitating Constructive Government-Industry Interations
+  container-title: Environmental Modelling & Software
+  container-title-short: Environmental Modelling & Software
+  DOI: 10.1016/j.envsoft.2003.08.018
+  ISSN: 1364-8152
+  issue: '7'
+  issued:
+    - year: 2004
+      month: 7
+      day: 1
+  language: en
+  page: 645-654
+  source: ScienceDirect
+  title: Visual Plumes mixing zone modeling software
+  type: article-journal
+  URL: https://www.sciencedirect.com/science/article/pii/S1364815203001890
+  volume: '19'
+
+- id: Jirka2004
+  accessed:
+    - year: 2021
+      month: 2
+      day: 22
+  author:
+    - family: Jirka
+      given: Gerhard H.
+  citation-key: Jirka2004
+  container-title: Environmental Fluid Mechanics
+  container-title-short: Environmental Fluid Mechanics
+  DOI: 10.1023/A:1025583110842
+  ISSN: 1573-1510
+  issue: '1'
+  issued:
+    - year: 2004
+      month: 3
+      day: 1
+  language: en
+  page: 1-56
+  source: Springer Link
+  title: >-
+    Integral Model for Turbulent Buoyant Jets in Unbounded Stratified Flows.
+    Part I: Single Round Jet
+  title-short: >-
+    Integral Model for Turbulent Buoyant Jets in Unbounded Stratified Flows.
+    Part I
+  type: article-journal
+  URL: https://doi.org/10.1023/A:1025583110842
+  volume: '4'
+
+- id: Lee2003
+  accessed:
+    - year: 2021
+      month: 2
+      day: 19
+  author:
+    - family: Lee
+      given: Joseph H. W.
+    - family: Chu
+      given: Vincent H.
+  citation-key: Lee2003
+  DOI: 10.1007/978-1-4615-0407-8
+  event-place: Boston, MA
+  ISBN: 978-1-4613-5061-3 978-1-4615-0407-8
+  issued:
+    - year: 2003
+  language: en
[…]
