### 2023-06-22 #5554 @pnsaevik
I wasn't aware of the citation syntax required in `paper.md`. Have fixed it now.
