### 2023-06-23 pnsaevik/effluent#25 @bastorer
Comparison with other models in paper
In `Statement of need`, please expand a little bit on if the underlying model methods and features are different compared with `Effluent`. That would align well with the `State of the Field` requirement from JOSS. That `Effluent` is open-source is great, but some discussion of the differences in functionality versus CORMIX, VISUAL PLUMES, and VISJET would be helpful.

[ This is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/5554 ]
