### 2023-06-23 pnsaevik/effluent#26 @bastorer
Error with CSV output. Dependencies?
I tried running the example given at the top of the documentation page under usage
```
(base) [14:13:32:Effluent:Ben]$ cat config.toml
# Characteristics of the pipe and the effluent flow
[pipe]
time = [1970-01-01]
flow = [0.2]
dens = [1000]
diam = [0.5]
depth = [100]
decline = [0]

# Characteristics of the ambient water masses
[ambient]
time = [1970-01-01]
depth = [100]
coflow = [[0]]
crossflow = [[0]]
dens = [[1000]]

# Output options
[output]
csv.file = "out.csv"
csv.float_format = "%.6g"
trajectory.step = 5           # Time between trajectory points [s]
trajectory.stop = 30          # Time of final trajectory point [s]
release.start = 1970-01-01    # Time of first release [date]
release.stop = 1970-01-01     # Time of final release [date]
release.step = 3600           # Time between releases [s]
```

but got an error when it tried to write the output file
```
(base) [14:12:58:Effluent:Ben]$ python
Python 3.9.13 (main, Aug 25 2022, 18:29:29)
[Clang 12.0.0 ] :: Anaconda, Inc. on darwin
>>> import effluent
>>> effluent.run('config.toml')
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/Ben/.local/lib/python3.9/site-packages/effluent/script.py", line 15, in run
    model.run()
  File "/Users/Ben/.local/lib/python3.9/site-packages/effluent/model.py", line 45, in run
    for _ in self.irun():
  File "/Users/Ben/.local/lib/python3.9/site-packages/effluent/model.py", line 60, in irun
    self.output.write(time, result)
  File "/Users/Ben/.local/lib/python3.9/site-packages/effluent/io.py", line 263, in write
    df.to_csv(
TypeError: to_csv() got an unexpected keyword argument 'lineterminator'
>>>
```

I switched to running in an isolated conda environment, and that seemed to resolve the issue, which makes me think it's a dependencies problem. Could you provide a list of dependencies (or, perhaps most ideally, an `environment.yml` file that can be used to build a working conda environment)? The `environment.yml` that I used is
```
name: Effluent

dependencies:
 - python>=3.7
 - netCDF4>=1.4
 - matplotlib>=3.1
 - pip
 - pip:
   - effluent
```
but it would be good to have an official one / confirmation about dependencies :-)

There's also a warning about precision conversions, that I figured I'd mention. The warning message includes a suggestion on how to resolve it, which might be nice to avoid warning-spamming :-)
```
UserWarning: Converting non-nanosecond precision datetime values to nanosecond precision. This behavior can eventually be relaxed in xarray, as it is an artifact from pandas which is now beginning to support non-nanosecond precision values. This warning is caused by passing non-nanosecond np.datetime64 or np.timedelta64 values to the DataArray or Variable constructor; it can be silenced by converting the values to nanosecond precision ahead of time.
  return self._dset.isel(time=0)
```

[This is in reference to JOSS review https://github.com/openjournals/joss-reviews/issues/5554]
