### 2023-06-23 pnsaevik/effluent#27 @bastorer
Providing output files for small tutorial problems
I have some of the example cases running, but it would be nice to be able to confirm that the outputs are as expected. When I run the first example under 'Usage' on the documentation, the output file is only 8 lines, and so an 'official' output for comparison could be included in the repository. For each of the examples listed on the documentation page, it would be nice if the expected output was available in the repository [ assuming that the files aren't onerously large ].

[This is in reference to JOSS review https://github.com/openjournals/joss-reviews/issues/5554]
