### 2023-06-23 pnsaevik/effluent#27 @bastorer
Sorry, just saw that the examples have the outputs in text on the documentation page. I'd recommend adding a `Tutorials` directory to the repository, which includes the `config.yml` and output files for each of those example cases, making it easy for new users to quickly run those cases and compare the outputs against expectations.
