### 2023-06-23 pnsaevik/effluent#26 @bastorer
I'm also running into issues with netcdf outputs, but think that might also be a dependencies issue. None of the provided examples use netcdf output, so to make sure that I'm not just doing something silly with the config file, could you provide an example that gives netcdf outputs?
