### 2023-06-23 pnsaevik/effluent#28 @bastorer
License mis-match
The paper mentions using the `Creative Commons Attribution 4.0 International` license, while the actual license file is the `MIT No Attribution License` (mostly - there's slight text difference, https://opensource.org/license/mit-0/). It doesn't look like CC BY 4 is OSI approved (at least, I can't find it in their approved list), but MIT-0 is.

Please unify the license choices with an OSI-approved license.

[This is part of JOSS review https://github.com/openjournals/joss-reviews/issues/5554]
