### 2023-06-23 pnsaevik/effluent#28 @bastorer
Ah, sorry, the 'CC BY 4' in the paper looks like that is the license _for the paper_, not the code. Please disregard.
