### 2023-06-27 pnsaevik/effluent#27 @pnsaevik
Hi @bastorer, thanks for taking the time to review this submission. I already have a directory like the one you are proposing, in `docs/source/examples`. Would it be enough to put a link to this directory in the readme file and the online documentation?
