### 2023-06-27 pnsaevik/effluent#26 @pnsaevik
Are you on Windows or Unix? I could provide an `environment.yml` file, absolutely, with some helpful text in the documentation about conda usage. On Windows, netCDF is tricky without proper conda configuration, but the unix integration server seems to run without problems while only using pip.
