### 2023-06-27 cfe1e6d
Make pytest fail on warnings
