+++ pytest.ini
@@ -4,3 +4,5 @@ minversion = 6.0
 addopts = -ra -q
 testpaths =
     tests
+filterwarnings =
+    error
