### 2023-06-27 pnsaevik/effluent#26 @pnsaevik
Your specific CSV problem is related to an API change in `pandas 1.5.0`. I can tell pip to only use newer versions, but first I'll check if I can make it compatible with older pandas versions as well.
