+++ pytest.ini
@@ -6,3 +6,4 @@ testpaths =
     tests
 filterwarnings =
     error
+    ignore:Converting non-nanosecond precision datetime values to nanosecond precision:UserWarning
