### 2023-06-27 8627d30
Use os-dependent line terminator in csv files
The lineterminator argument to to_csv was not backwards-compatible,
and was unnecessary anyway.
