### 2023-06-27 pnsaevik/effluent#26 @pnsaevik
The csv problem should be resolved by #30. Can you confirm?
