### 2023-06-28 pnsaevik/effluent#27 @bastorer
That's great! Yeah, just adding some text in the readme and documentation mentioning that the .toml and output files for the example cases are available in `docs/sources/examples` would be perfect.
