### 2023-06-28 pnsaevik/effluent#26 @bastorer
> Are you on Windows or Unix?
Unix [more specifically Mac/OS X]

> I could provide an environment.yml file, absolutely, with some helpful text in the documentation about conda usage.
That would be great!
