### 2023-06-28 pnsaevik/effluent#26 @bastorer
> The csv problem should be resolved by #30. Can you confirm?

Unfortunately, the problem seems to persist. I've uploaded the output of `pip install --verbose effluent` here
[install.log](https://github.com/pnsaevik/effluent/files/11899646/install.log).

```
(base) [16:15:40:Effluent:Ben]$ python
Python 3.9.13 (main, Aug 25 2022, 18:29:29)
[Clang 12.0.0 ] :: Anaconda, Inc. on darwin
>>> import effluent
>>> effluent.run('config.toml')
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/Ben/opt/anaconda3/lib/python3.9/site-packages/effluent/script.py", line 15, in run
    model.run()
  File "/Users/Ben/opt/anaconda3/lib/python3.9/site-packages/effluent/model.py", line 45, in run
    for _ in self.irun():
  File "/Users/Ben/opt/anaconda3/lib/python3.9/site-packages/effluent/model.py", line 60, in irun
    self.output.write(time, result)
  File "/Users/Ben/opt/anaconda3/lib/python3.9/site-packages/effluent/io.py", line 263, in write
    df.to_csv(
TypeError: to_csv() got an unexpected keyword argument 'lineterminator'
```
