### 2023-06-29 pnsaevik/effluent#26 @pnsaevik
I haven't yet created a new release number for the update, so it's still on 1.2. Which means you'll have to run pip uninstall effluent first, otherwise pip won't download the new version. And use pip git+https://github.com/pnsaevik/effluent to be sure to download the new version ("pip install effluent" installs from PyPI).

I'm travelling in June, but will return to this review in August. See you until then.
