### 2023-08-01 #5554 @bastorer
Hi @kthyng . @pnsaevik mentioned being on holidays until August (in this issue https://github.com/pnsaevik/effluent/issues/26). I asked them to send me a ping where they're back and ready to resume :-)
