### 2023-08-01 #5554 @kthyng
@bastorer Ah, thanks. Summer is pretty spotty for a lot of us!
