### 2023-08-04 89a74c1
Add reference to worked examples
Add reference to worked examples in the readme file, and in the front
page of the documentation.
