+++ docs/source/examples.rst
@@ -1,3 +1,5 @@
+.. _examples_page:
+
 =============
 Examples
 =============
