### 2023-08-04 pnsaevik/effluent#26 @pnsaevik
I'm back now. Can you test whether the development version solves the csv issue? To install, do the following:
1. Uninstall effluent, `pip uninstall effluent`
2. Install the development version, `pip install git+https://github.com/pnsaevik/effluent`
