### 2023-08-14 pnsaevik/effluent#26 @castelao
@pnsaevik , I started a new environment and re-installed the latest version of effluent. The tests pass and that example worked, but the nanoseconds warning is still complaining.

For reference:
effluent   097c01064

python                    3.11.4
pandas                    2.0.3
