### 2023-08-14 pnsaevik/effluent#32 @castelao
Community guidelines
Hi @pnsaevik , I hope you had a good summer break. I couldn't find the 'Community guidelines', which is a requirement for JOSS.

[This is in reference to JOSS review https://github.com/https://github.com/openjournals/joss-reviews/issues/5554]
