### 2023-08-15 pnsaevik/effluent#26 @pnsaevik
Strange, I cannot reproduce the warning, neither on windows or linux, nor on circle ci. I use the same python and pandas versions. Anyway, since it's just a warning, I suggest we close this issue now.
