### 2023-08-15 pnsaevik/effluent#32 @pnsaevik
I've now added a very brief (but hopefully sufficient) statement in #33. I don't think there is need for anything more verbose, do you agree?
