### 2023-08-15 pnsaevik/effluent#25 @pnsaevik
Can you have a look at pull request #34 and see if it is sufficient? I don't think a detailed comparison is warranted here (after all, it's done in a published paper already..), but I've added a few key highlights.
