+++ docs/source/credits.rst
@@ -1,0 +1,10 @@
+Credits
+=======
+
+This package is funded and developed by the
+`Institute of Marine Research <https://www.hi.no>`_ (Norway) in support for
+various research projects related to effluent wastewater and mine tailings.
+
+Development lead
+----------------
+* Pål Næverlid Sævik <paal.naeverlid.saevik@hi.no>
