### 2023-08-27 #5554 @castelao
For the record, I made a few comments and suggestions to @pnsaevik early in the summer (BTW @pnsaevik was very positive and proactive, even for modifications that wouldn't be strictly required), but I forgot to add links to here. Those modifications and discussions can be found as closed PRs in the [effluent repository](https://github.com/pnsaevik/effluent/pulls?q=is%3Apr+is%3Aclosed).
