### 2023-08-28 pnsaevik/effluent#26 @bastorer
Hey @pnsaevik , I'm happy to confirm that the test passed without errors or warnings on my end :-)

Feel free to close the issue!
