### 2023-08-28 pnsaevik/effluent#25 @bastorer
@pnsaevik the new citation and extra description are nice additions. Let me know when you merge #34 , I'll give the pdf one last look-over, but otherwise feel free to close this issue with PR #34
