### 2023-08-28 pnsaevik/effluent#35 @bastorer
Output from tests
Are the tests in the `/tests` directory expected to output anything? If I run, e.g. `python test_solver.py`, it runs happily with no output.

If those are tests that intended for the user to verify installation, I'd recommend adding (at least) two print statements to each. One that prints at the beginning, saying that the test is starting, and one that prints at the end, saying if the test passed or failed.

Also, if you do decide to include a conda `.yml` file for building an Effluent environment, you'll also want to include `pytest`, which is necessary for these scripts to run.

[This is part of JOSS review https://github.com/https://github.com/openjournals/joss-reviews/issues/5554]
