### 2023-08-28 pnsaevik/effluent#36 @bastorer
Error writing release time to netcdf output
When outputting to netcdf files, I receive the error below (using a freshly pip-installed version via `pip install git+https://github.com/pnsaevik/effluent`)

The other variables have `dtypes` of `float64`, but `relase_time` has `dtype` of `datetime64[ns]`, which isn't supported with netcdf. You might to add a check here to cast `release_time` to another type (double, or long unsigned int, etc). Using an 'add_offset' attribute for that variable might be helpful, but not required.

```
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/Ben/opt/anaconda3/envs/Effluent/lib/python3.11/site-packages/effluent/script.py", line 15, in run
    model.run()
  File "/Users/Ben/opt/anaconda3/envs/Effluent/lib/python3.11/site-packages/effluent/model.py", line 45, in run
    for _ in self.irun():
  File "/Users/Ben/opt/anaconda3/envs/Effluent/lib/python3.11/site-packages/effluent/model.py", line 60, in irun
    self.output.write(time, result)
  File "/Users/Ben/opt/anaconda3/envs/Effluent/lib/python3.11/site-packages/effluent/io.py", line 337, in write
    write_xr_to_nc(result, self.dset)
  File "/Users/Ben/opt/anaconda3/envs/Effluent/lib/python3.11/site-packages/effluent/io.py", line 396, in write_xr_to_nc
    nc_var = nc_dset.createVariable(
             ^^^^^^^^^^^^^^^^^^^^^^^
  File "src/netCDF4/_netCDF4.pyx", line 2962, in netCDF4._netCDF4.Dataset.createVariable
  File "src/netCDF4/_netCDF4.pyx", line 4150, in netCDF4._netCDF4.Variable.__init__
TypeError: illegal primitive data type, must be one of dict_keys(['S1', 'i1', 'u1', 'i2', 'u2', 'i4', 'u4', 'i8', 'u8', 'f4', 'f8']), got datetime64[ns]
```

[This is part of JOSS review https://github.com/https://github.com/openjournals/joss-reviews/issues/5554]
