### 2023-08-30 pnsaevik/effluent#35 @pnsaevik
There are probably many ways to do these things. In `effluent`, I have followed the fairly standard approach of separating the test suite from the main package, so regular users don't have access to the tests. Therefore, pytest is not a part of the package requirements. To set up a development environment one should clone the repo using git, and install pytest in addition to the package itself. To run the tests, cd into the repo dir and run `pytest` with no arguments. The tests should then run (54 of them). The command `python test_solver.py` does not actually run anything.

In other words, something like this:
```
cd name_of_dev_dir
git clone https://github.com/pnsaevik/effluent .
conda create -n effluent python xarray netcdf4 holoviews scipy dask
conda activate effluent
pip install pytest
pip install -e .
pytest
```

The conda environment is optional, but may be helpful on Windows. See also the "contribute code" entry from the documentation at https://github.com/pnsaevik/effluent/blob/main/docs/source/contributing.rst
