### 2023-08-30 3351d16
Create example with netcdf output
