### 2023-08-30 b8cb884
Add datetime conversion to write_xr_to_nc
