### 2023-08-30 2a680c2
Add datetime conversion to append_xr_to_nc
