### 2023-08-30 6bd13b0
Allow custom specification of date units in netcdf output
