### 2023-08-30 df7dfaa
Use netcdf output in roms example
