### 2023-08-30 96b343a
Only check closeness match for netcdf files
