### 2023-08-30 pnsaevik/effluent#36 @pnsaevik
This turned out to be a bug, thanks for the report. Apparently I had not tested the netcdf output feature properly. Can you confirm that PR #37 closes the issue?
