### 2023-09-05 pnsaevik/effluent#38 @castelao
Missing new reference
I believe that `paper.bib` is missing the new reference `Palomar2012`.

Look at the latest (compiled paper)[https://github.com/openjournals/joss-papers/blob/joss.05554/joss.05554/10.21105.joss.05554.pdf]

[This is in reference to JOSS review https://github.com/https://github.com/https://github.com/openjournals/joss-reviews/issues/5554]
