### 2023-09-05 pnsaevik/effluent#39 @castelao
Comparison with previous work
A follow-up on #25.

@pnsaevik , I think #34 was a great addition to the paper. Did I understand correctly that you discuss only the comparison with VISJET, but leave the discussion on CORMIX and VISUAL PLUMES to Palomar 2012? If so, it makes sense to me not to repeat Palomar 2012, but do you think it is possible to summarize in 1-2 lines the differences of effluent with those other two? As you mention, being Open Source would be argument enough for me to rather use `effluent`, but are there special cases or situations that the user should be aware of? Are there different assumptions between the three of you?

[This is in reference to JOSS review https://github.com/https://github.com/https://github.com/openjournals/joss-reviews/issues/5554]
