### 2023-09-06 d74a13f
Add figure to the paper
