+++ docs/paper.md
@@ -33,6 +33,8 @@ dilution process may help discover this type of problems before they appear,
 and can guide the design of an outfall system to minimize environmental
 impact.
 
+![Wastewater dilution and rise, as computed by the package](paper-figure.png){ width=50% }
+
 # Statement of need
 
 `effluent` is an open-source python package for simulating the dispersion of
