### 2023-09-06 484ea4e
Adding conda environment files both for regular usage, as well as for running the test suite. Updated the README accordingly. NOTE: The yml file currently points to the git repo       when installing, instead of pip. Update the yml       when the pip version is updated with the latest changes.
