+++ tests/test_environment.yml
@@ -1,0 +1,10 @@
+name: EffluentPyTest
+
+dependencies:
+ - python>=3.7
+ - netCDF4>=1.4
+ - matplotlib>=3.1
+ - pytest
+ - pip
+ - pip:
+   - git+https://github.com/pnsaevik/effluent
