+++ environment.yml
@@ -1,0 +1,9 @@
+name: Effluent
+
+dependencies:
+ - python>=3.7
+ - netCDF4>=1.4
+ - matplotlib>=3.1
+ - pip
+ - pip:
+   - git+https://github.com/pnsaevik/effluent
