### 2023-09-07 abc6a80
Increase figure width
