### 2023-09-07 pnsaevik/effluent#39 @pnsaevik
I've added some lines in #43, but I'm afraid of "bloating" the paper too much. So I'm closing this issue for now.
