### 2023-09-07 3147fc9
Create conda environment and update development guidance
