+++ environment.yml
@@ -1,0 +1,12 @@
+name: effluent
+
+dependencies:
+ - python>=3.7
+ - netCDF4>=1.4
+ - xarray
+ - pandas
+ - scipy
+ - dask
+ - pip
+ - pip:
+   - effluent
