### 2023-09-08 #5554 @bastorer
@kthyng and @pnsaevik The updated pdf is nice, and @pnsaevik has responded to all of my tickets/issues. I'm happy to declare my checklist complete and congratulate @pnsaevik on the nice user-friendly (and open source!) package that they've put together :-)
