### 2023-09-11 27f35a2
Add bibtex entry for palomar paper.bib
