+++ docs/paper.bib
@@ -62,7 +62,20 @@
 	author = {Shchepetkin, Alexander F. and McWilliams, James C.},
 	month = jan,
 	year = {2005},
-	note = {tex.ids: Shchepetkin2005a, Shchepetkin2005b},
 	pages = {347--404},
-	file = {ScienceDirect Full Text PDF:C\:\\Users\\a5606\\Documents\\Private\\Literature\\Zotero\\storage\\74KFE6GG\\Shchepetkin and McWilliams - 2005 - The regional oceanic modeling system (ROMS) a spl.pdf:application/pdf;ScienceDirect Full Text PDF:C\:\\Users\\a5606\\Documents\\Private\\Literature\\Zotero\\storage\\TX5N74AN\\Shchepetkin and McWilliams - 2005 - The regional oceanic modeling system (ROMS) a spl.pdf:application/pdf;ScienceDirect Snapshot:C\:\\Users\\a5606\\Documents\\Private\\Literature\\Zotero\\storage\\CPB7DVS8\\S1463500304000484.html:text/html;ScienceDirect Snapshot:C\:\\Users\\a5606\\Documents\\Private\\Literature\\Zotero\\storage\\ZHITQ7MY\\S1463500304000484.html:text/html},
+}
+
+@article{palomar:2012,
+  title = {Near Field Brine Discharge Modelling Part 1: {{Analysis}} of Commercial Tools},
+  shorttitle = {Near Field Brine Discharge Modelling Part 1},
+  author = {Palomar, P. and Lara, J. L. and Losada, I. J. and Rodrigo, M. and Alv{\'a}rez, A.},
+  year = {2012},
+  month = mar,
+  journal = {Desalination},
+  volume = {290},
+  pages = {14--27},
+  issn = {0011-9164},
+  doi = {10.1016/j.desal.2011.11.037},
+  urldate = {2023-09-06},
+  keywords = {Brine,Discharge,Environment,Modelling,Near field,Validation},
 }
