+++ CITATION.cff
@@ -5,6 +5,7 @@ authors:
   given-names: "Pål Næverlid"
   orcid: "https://orcid.org/0000-0002-7301-2008"
 title: "Effluent: A Python package for modelling effluent discharge"
+doi: 10.5281/zenodo.8333370
 version: 1.3
 date-released: 2023-09-11
 license: MIT
