### 2023-09-11 #5554 @pnsaevik
Hi @kthyng,

The Zenodo doi is `10.5281/zenodo.8333370`. The license and title was added automatically from GitHub, but it was wrong, so I had to change it manually. Zenodo still shows the auto-generated title when I access the record without being logged in, but hopefully it will resolve itself when the cache refreshes eventually.

The official version number with all the latest changes incorporated, is 1.3. It's also published on PyPI here: https://pypi.org/project/effluent/

The most recent version of the documentation is also online, at https://effluent.readthedocs.io/
