### 2023-09-11 #5554 @castelao
@pnsaevik , you can control Zenodo's DOI metadata with a [.zenodo.json](https://github.com/castelao/inception) file. Here [an example](https://github.com/castelao/GSW-rs/blob/main/.zenodo.json). Note that this would only affect the next release.
