### 2023-09-13 de7e9cd
Remove brackets
