+++ src/effluent/__init__.py
@@ -5,7 +5,7 @@ main entry point of the application.
 
 from .script import run
 
-__version__ = "1.3"
+__version__ = "1.3.1"
 """
 The version number
 """
