### 2023-09-13 6d2251e
Remove faulty zenodo metadata entry
