+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.2] - 2023-09-13
+
+### Fixed
+- Remove faulty zenodo metadata entry
+
+
 ## [1.3.1] - 2023-09-13
 
 ### Added
