### 2023-09-13 #5554 @pnsaevik
> A comment on your paper @pnsaevik: The citation on line 14 should be inline instead of parenthetical.

Thanks. I've fixed the issue in v1.3.2 (and added a zenodo metadata file as well). But it's probably not worth re-iterating the process since nothing else is changed.
